requirements.txt
fastapi
uvicorn
fpdf
pillow
torch
torchvision
python-multipart
requests